+++ qldpc/circuits.py
@@ -22,11 +22,10 @@ import functools
 from collections.abc import Callable, Collection, Sequence
 
 import numpy as np
-import numpy.typing as npt
 import stim
 
 from qldpc import abstract, codes
-from qldpc.objects import Pauli
+from qldpc.objects import Pauli, op_to_string
 
 
 def restrict_to_qubits(func: Callable[..., stim.Circuit]) -> Callable[..., stim.Circuit]:
@@ -39,28 +38,6 @@ def restrict_to_qubits(func: Callable[..., stim.Circuit]) -> Callable[..., stim.
         return func(*args, **kwargs)
 
     return qubit_func
-
-
-def op_to_string(op: npt.NDArray[np.int_], flip_xz: bool = False) -> stim.PauliString:
-    """Convert an integer array that represents a Pauli string into a stim.PauliString.
-
-    The (first, second) half the array indicates the support of (X, Z) Paulis, unless flip_xz==True.
-    """
-    assert len(op) % 2 == 0
-    support_xz = np.array(op, dtype=int).reshape(2, -1)
-    if flip_xz:
-        support_xz = support_xz[::-1, :]
-    paulis = [Pauli((support_xz[0, qq], support_xz[1, qq])) for qq in range(support_xz.shape[1])]
-    return stim.PauliString(map(str, paulis))
-
-    num_qubits = len(op) // 2
-    paulis = ""
-    for qubit in range(num_qubits):
-        val_x = int(op[qubit])
-        val_z = int(op[qubit + num_qubits])
-        pauli = Pauli((val_x, val_z))
-        paulis += str(pauli if not flip_xz else ~pauli)
-    return stim.PauliString(paulis)
 
 
 @restrict_to_qubits
